Tidy up config classes
